backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
python-dotenv==1.2.1
pydantic==2.12.5
groq==0.37.1
langchain-groq==1.1.1
json-repair==0.55.0
PyMuPDF==1.26.7
pillow==12.1.0
openpyxl==3.1.5